Refactor tests
- Test class for each group of Ops
- Disallow parallel test execution
