+++ build.sbt
@@ -4,6 +4,8 @@ version := "1.0"
 
 scalaVersion := "2.11.7"
 
+parallelExecution in Test := false
+
 libraryDependencies ++= Seq(
   "net.debasishg" %% "redisclient" % "3.0",
   "org.scalaz" %% "scalaz-core" % "7.1.4",
